[feat] 사용자는 회원 탈퇴를 할 수 있다 (#21)
* fix: 공백 제거(#12)

* fix: 메인페이지 head 태그 추가(#12)

* fix: 공백 추가(#12)

* feat: 로그인 성공 시 깃허브 토큰 redis에 저장하도록 추가(#12)

* fix: AuthenticationSuccessHandler에 의존성 주입을 위한 매개변수 추가(#12)

* feat: 사용자 탈퇴 기능(#12)

* feat: HTTP method 확장을 위한 HiddenHttpMethodFilter 설정 추가(#12)

* fix: 로그인 성공 화면 title 수정(#12)

* feat: 마이페이지 탈퇴 화면 추가(#12)

* fix: githubId를 int 타입으로 수정(#12)

* feat: GitHub 토큰 해제를 위한 WebClient Bean 등록(#12)

* fix: GitHub revoke 로직을 RestTemplate에서 WebClient로 전환(#12)

* feat: WebClient 기반 GitHub access token 해제 기능 추가(#12)

* fix: redis에서 github access token을 불러올 때 delete에서 unlink 메서드로 수정(#12)

* fix: 사용하지 않는 CustomOAuth2UserService 필드 제거(#12)

* fix: 사용하지 않는 application.yml 제거(#12)

* chore: application-test.yml에 테스트용 더미 데이터 추가(#12)

* test: WebClient MockitoBean으로 주입되도록 수정(#12)

* feat: EMPTY_OAUTH_TOKEN 예외 작성(#12)

* fix: GitHub ID 추출 시 불필요한 형 변환 제거(#12)

* fix: access token null/blank 검사에 StringUtils.hasText 사용 및 GeneralException으로 변경(#12)

* fix: Optional 제거(#12)

+++ src/main/java/com/specialwarriors/conal/user/domain/User.java
@@ -24,7 +24,7 @@ public class User {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private long githubId;
+    private int githubId;
 
     @Column(nullable = false)
     private String username;
@@ -35,7 +35,7 @@ public class User {
     @OneToMany(mappedBy = "user")
     private List<GithubRepo> githubRepos = new ArrayList<>();
 
-    public User(long githubId, String username, String avatarUrl) {
+    public User(int githubId, String username, String avatarUrl) {
         this.githubId = githubId;
         this.username = username;
         this.avatarUrl = avatarUrl;
